app/requirements.txt
streamlit==1.51.0
bcrypt==5.0.0
PyYAML==6.0.3
streamlit-authenticator==0.3.3